refactor: Split middleware.py into middleware/ directory
Split the 1274-line middleware.py into a modular directory structure:

arbiter_ai/core/middleware/
├── __init__.py          # Public exports
├── base.py              # Middleware ABC, MiddlewarePipeline
├── cache_backends.py    # CacheBackend, MemoryCacheBackend, RedisCacheBackend
├── caching.py           # CachingMiddleware
├── logging_mw.py        # LoggingMiddleware
├── metrics.py           # MetricsMiddleware
├── rate_limiting.py     # RateLimitingMiddleware
└── monitor.py           # monitor() context manager

Benefits:
- Single responsibility per file
- Easier to test individual middleware
- Clearer import structure
- Easier to add new middleware types

No breaking changes to public API - all existing imports work.

Closes #109
